autograder_website/backend/requirements.txt
flask
flask_cors
psycopg2-binary
python-dotenv
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy